Bumping version to 2.0.4-SNAPSHOT and fixing license URL
There is already a 2.0.3 tag so to be able to branch off properly
from the origin's tags and releases bumping version to 2.0.4 to
be able to make a clean branch point, to avoid version 2.0.3 to
mean to separate things.

Also fixing the license URL to be the correct one for the MIT
license.